structuure change 5
