+++ problems/binary_search/essential/ez__binary_search.py
