+++ solutions.py
@@ -169,72 +169,7 @@ class LongestCommonPrefix:
         return shortest_word
 
 
-class MergeTwoSortedLists:
-    @staticmethod
-    def my_solution(l1: ListNode, l2: ListNode) -> ListNode:
-        """
-        You are given the heads of two sorted linked lists list1 and list2.
-        Merge the two lists in a one sorted list. The list should be made by splicing together the nodes of the first two lists.
-        Return the head of the merged linked list.
 
-        Input: list1 = [1,2,4], list2 = [1,3,4]
-        Output: [1,1,2,3,4,4]
-        """
-        dummy = ListNode()
-        temp = dummy
-
-        while l1 and l2:
-            if l1.val < l2.val:
-                temp.next = l1
-                l1 = l1.next
-            else:
-                temp.next = l2
-                l2 = l2.next
-
-            temp = temp.next
-            
-        if l1:
-            temp.next = l1
-        elif l2:
-            temp.next = l2
-
-        return transfer_linked_nodes_to_list(dummy.next) # if run on leetcode server, then simply return dummy.next
 
         
-
-
-
-
-class SameTree:
-    """
-    https://leetcode.com/problems/same-tree/
-    Given the roots of two binary trees p and q, write a function to check if they are the same or not.
-    Two binary trees are considered the same if they are structurally identical, and the nodes have the same value.
-
-    Input: p = [1,2,3], q = [1,2,3]
-    Output: true
-    """
-
-    @staticmethod
-    def my_solution(p: BinaryTreeNode, q: BinaryTreeNode) -> bool:
-        """
-        tech use : Depth First Search (DFS)
-        https://www.techiedelight.com/depth-first-search/
-
-        https://www.youtube.com/watch?time_continue=453&v=vRbbcKXCxOw&feature=emb_title
-        """
-
-        def _is_same_tree(p, q):
-            if not p and not q: 
-                # if both p and q are None (empty tree)
-                return True
-            
-            if not p or not q or p.val != q.val:
-                # if only one of p or q is empty or the value for both are different
-                return False
-        
-            return (_is_same_tree(p.left, q.left) and 
-                    _is_same_tree(p.right, q.right))
-        
-        return (_is_same_tree(p, q))
